3.7.2.2 =======
Wavenumber computation & better integration with ORCS
-----------------------------------------------------

The whole spectrum computation process can now be done in wavenumber
(useful to avoid the mutiple interpolation nescessary to move from a
regular wavenumber space to an iregular wavelength space back and
forth).

It is also possible to compute an uncalibrated spectrum. This way
there is absolutly no interpolation made during the spectrum
computation. The output can be used by ORCS directly and ORCS itself
does not have any interpolation to do for the extraction of the lines
parameters. **This ensure that the spectral information is not distorded
at all during the process**.

The filter correction during the calibration process takes into
account the fact that no wavelength calibration has been done.

* Important modified methods:

  * :py:meth:`~process.Interferogram.compute_spectrum`
  * :py:meth:`~process.Spectrum.calibrate`

* New keyword added in the option file: WAVENUMBER, WAVE_CALIB

Miscellaneous
-------------

* :py:meth:`~orbs.Orbs.__init__` simplified by the use of
  :py:class:`orb.core.OptionFile` previously used only by
  :py:class:`orcs.orcs.SpectralCube()`

* doc updated

ORB's scripts
-------------

* move ORB's scripts (dstack, combine, rollxz, rollyz, reduce) from
  orbs/scripts to orb/scripts so that only ORBS specific scripts are
  in orbs/scripts.

* create **unstack** script to unstack a cube into a set of frames

	modifié:         doc/changelog.rst
	modifié:         orbs/orbs.py
	modifié:         orbs/process.py
	modifié:         orbs/version.py
	supprimé:        scripts/combine
	supprimé:        scripts/dstack
	supprimé:        scripts/reduce
	supprimé:        scripts/rollxz
	supprimé:        scripts/rollyz